requirements.txt
sentence-transformers
faiss-cpu
transformers
datasets
torch
scikit-learn
python-dotenv
streamlit
numpy
tqdm
spacy
matplotlib
pandas
joblib
huggingface_hub
scipy
